% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUT.R
\name{as_NUT}
\alias{as_NUT}
\title{#' @method vec_cast.hd_NUT logical
#' @export
vec_cast.hd_NUT.logical <- function(x, to, ...) NUT(x)}
\usage{
as_NUT(x)
}
\description{
#' @method vec_cast.logical hd_NUT
#' @export
vec_cast.logical.hd_NUT <- function(x, to, ...) NUT(x)
}
