% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKT.R
\name{as_UKT}
\alias{as_UKT}
\title{#' @method vec_cast.hd_UKT logical
#' @export
vec_cast.hd_UKT.logical <- function(x, to, ...) UKT(x)}
\usage{
as_UKT(x)
}
\description{
#' @method vec_cast.logical hd_UKT
#' @export
vec_cast.logical.hd_UKT <- function(x, to, ...) UKT(x)
}
