% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_hdtype.R
\name{guess_hdtype}
\alias{guess_hdtype}
\title{hdtype}
\usage{
guess_hdtype(v)
}
\arguments{
\item{v}{a value for which you want to evaluate}
}
\value{
An available_hdtype value
}
\description{
evaluates which of the available_hdtypes is the object
}
\examples{

# Num hdtype
value <- c("1",NA,"2")
guess_hdtype(value)

# Cat hdtype
value <- c("MSPS-CD-166-2020", "003-2020", "0811 - 2020")
guess_hdtype(value)

# Pct hdtype
value <- c(0.3, 0.4, 1)
guess_hdtype(value)


}
