% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdTypes-package.R
\docType{package}
\name{hdtype-package}
\alias{hdtype}
\alias{hdtype-package}
\title{hdtype: Homodatum type definition}
\description{
Vector definitions for abstract data types
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datasketch/hdtype}
  \item Report bugs at \url{https://github.com/datasketch/hdtype/issues}
}

}
\author{
\strong{Maintainer}: Juan Pablo Marin Diaz \email{jpmarindiaz@gmail.com}

}
\keyword{internal}
