% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdTypes-package.R
\docType{package}
\name{hdTypes-package}
\alias{hdTypes}
\alias{hdTypes-package}
\title{hdTypes: Homodatum types}
\description{
Vector definitions for abstract data types
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datasketch/hdTypes}
  \item Report bugs at \url{https://github.com/datasketch/hdTypes/issues}
}

}
\author{
\strong{Maintainer}: Juan Pablo Marin Diaz \email{jpmarindiaz@gmail.com}

}
\keyword{internal}
