% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdtype.R
\name{is_any_hdtype}
\alias{is_any_hdtype}
\title{hdtype Vectors}
\usage{
is_any_hdtype(x)
}
\arguments{
\item{x}{hdtype object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is one of the \code{\link[=available_hdtypes]{available_hdtypes()}} or not.
}
\description{
test for objects of type "hdtype"
}
\examples{

some_cat_value <- Cat("value1")
is_any_hdtype(some_cat_value)

}
