% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clr.R
\name{is_Clr}
\alias{is_Clr}
\title{Clregory Vectors}
\usage{
is_Clr(x)
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is of type hd_Clr or not.
}
\description{
test for objects of type "hd_Clr"
}
\examples{
some_value <- Clr(c("Clr1", "Clr2"))
is_Clr(some_value)

}
