% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdType.R
\name{is_hdtype}
\alias{is_hdtype}
\title{hdtype Vectors}
\usage{
is_hdtype(x)
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is of type hdtype or not.
}
\description{
test for objects of type "hdtype"
}
\examples{

some_value <- hdtype("Cat")
is_hdtype(some_value)

}
