% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdvector.R
\name{is_hdvector}
\alias{is_hdvector}
\title{Check if x is an hdvector}
\usage{
is_hdvector(x)
}
\arguments{
\item{x}{hdtype object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is one of the \code{\link[=available_hdtypes]{available_hdtypes()}} or not.
}
\description{
test for objects of type "hdvector"
}
\examples{

x <- Cat("value1")
is_hdvector(x)

}
