% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUT.R
\name{is_NUT}
\alias{is_NUT}
\title{Category Vectors}
\usage{
is_NUT(x)
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is of type hd_NUT or not.
}
\description{
test for objects of type "hd_NUT"
}
\examples{
some_value <- Cat(c("Cat1", "Cat2"))
is_Cat(some_value)

}
