% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKT.R
\name{is_UKT}
\alias{is_UKT}
\title{Category Vectors}
\usage{
is_UKT(x)
}
\arguments{
\item{x}{object to be coerced or tested}
}
\value{
returns TRUE or FALSE depending on whether its argument is of type hd_UKT or not.
}
\description{
test for objects of type "hd_UKT"
}
\examples{
some_value <- Cat(c("Cat1", "Cat2"))
is_Cat(some_value)

}
