% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUT.R
\name{NUT}
\alias{NUT}
\title{Null Type Vectors}
\usage{
NUT(x = character(), format = NULL)
}
\arguments{
\item{x}{object to be created as NUT type}
}
\description{
Creates objects of type "hd_NUT". Usually for NA vectors
}
\examples{
x <- c(NA, NA)
ukts <- new_NUT(x)
class(ukts)

}
