% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UKT.R
\name{UKT}
\alias{UKT}
\title{UnknownType Vectors}
\usage{
UKT(x = character(), format = NULL)
}
\arguments{
\item{x}{object to be created as UKT type}
}
\description{
Creates objects of type "hd_UKT". Usually for NA vectors
}
\examples{
x <- c(NA, NA)
ukts <- new_UKT(x)
class(ukts)

}
