% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdType.R
\name{which_hdtype}
\alias{which_hdtype}
\title{hdtype Vectors}
\usage{
which_hdtype(x)
}
\arguments{
\item{x}{hdtype object to be coerced or tested}
}
\value{
returns the name of the hdtype value. You can see the valid hdtypes with \code{\link[=available_hdtypes]{available_hdtypes()}}
}
\description{
Detect which hdtype is the value.
}
\examples{

some_cat_value <- Cat("value1")
which_hdtype(some_cat_value)

}
